Use normal mode for vim-fugitive summary commands
There is a quirky state that's triggered after closing a buffer from RPC
while in INSERT, which is that you switch to the previous buffer in
INSERT mode. So if I close the commit message window while in INSERT
mode, I auto switch back to the vim-fugitive summary buffer in the
same mode, which you can't usually enter. This causes all of the
commands to fail, until you switch back to normal mode. So my solution
to avoid this is to is automate switching.